+++ ThirdPartyAdapters/unity/unity/src/test/kotlin/com/google/ads/mediation/unity/UnityMediationBannerAdTest.kt
@@ -51,7 +51,7 @@ class UnityMediationBannerAdTest {
   @Before
   fun setUp() {
     val unityBannerSize: UnityBannerSize? =
-      UnityAdsAdapterUtils.getUnityBannerSize(activity, adSize)
+      UnityAdsAdapterUtils.getUnityBannerSize(activity, adSize, /* isRtb= */ false)
     bannerView = BannerView(activity, TEST_PLACEMENT_ID, unityBannerSize)
     unityMediationBannerAd =
       UnityMediationBannerAd(